// frontend/src/api/generated/models/PendingEventDTO.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type PendingEventDTO = {
    draft_id: number;
    title: string;
    announced_at?: (string | null);
    source?: (string | null);
    content?: (string | null);
    source_url?: (string | null);
    importance_hint?: (number | null);
    ai_suggestions?: (Record<string, any> | null);
};